break down json schema form types
