+++ src/components/SmartContractJsonSchema/index.tsx
@@ -5,7 +5,7 @@ import { parse, stringify } from "lossless-json";
 import { usePrevious } from "@/hooks/usePrevious";
 import { useStore } from "@/store/useStore";
 
-import { DereferencedSchemaType } from "@/constants/jsonSchema";
+import { type DereferencedSchemaType } from "@/constants/jsonSchema";
 
 import { dereferenceSchema } from "@/helpers/dereferenceSchema";
 import { getTxnToSimulate } from "@/helpers/sorobanUtils";
@@ -19,8 +19,7 @@ import { ErrorText } from "@/components/ErrorText";
 
 import { AnyObject, SorobanInvokeValue } from "@/types/types";
 
-import { JsonSchemaFormRenderer } from "./JsonSchemaFormRenderer";
-import { JsonSchemaFormRendererTwo } from "./JsonSchemaFormRendererTwo";
+import { JsonSchemaRenderer } from "./JsonSchemaRenderer";
 
 export const JsonSchemaForm = ({
   name,
@@ -51,9 +50,6 @@ export const JsonSchemaForm = ({
     funcSchema,
     name,
   );
-
-  // console.log("[JsonSchemaForm] funcSchema: ", funcSchema);
-  // console.log("[JsonSchemaForm] dereferencedSchema: ", dereferencedSchema);
 
   const requiredFields = dereferencedSchema.required;
   const [formError, setFormError] = useState<AnyObject>({});
@@ -152,19 +148,7 @@ export const JsonSchemaForm = ({
         {renderTitle(name, description)}
 
         <Box gap="md" key={name}>
-          {/* <JsonSchemaFormRenderer
-            name={name}
-            schema={dereferencedSchema as JSONSchema7}
-            onChange={onChange}
-            formError={formError}
-            setFormError={setFormError}
-            parsedSorobanOperation={
-              parse(
-                sorobanOperation.params.invoke_contract,
-              ) as SorobanInvokeValue
-            }
-          /> */}
-          <JsonSchemaFormRendererTwo
+          <JsonSchemaRenderer
             name={name}
             schema={dereferencedSchema as JSONSchema7}
             onChange={onChange}
