+++ src/components/SmartContractJsonSchema/renderArrayType.tsx
@@ -1,23 +1,17 @@
 import type { JSONSchema7 } from "json-schema";
+import { Button, Card, Text } from "@stellar/design-system";
+import { get } from "lodash";
 
 import { jsonSchema } from "@/helpers/jsonSchema";
+
+import { Box } from "@/components/layout/Box";
+import { LabelHeading } from "@/components/LabelHeading";
 
 import type {
   AnyObject,
   JsonSchemaFormProps,
   SorobanInvokeValue,
 } from "@/types/types";
-import { Box } from "../layout/Box";
-import { LabelHeading } from "../LabelHeading";
-import { Button, Card, Text } from "@stellar/design-system";
-
-import { get } from "lodash";
-
-// @TODO
-// Tuple Type (scSpecTypeTuple)
-//   if (jsonSchema.isTuple(schema)) {
-//     return <div>isTuple</div>;
-//   }
 
 export const renderArrayType = ({
   schema,
@@ -32,7 +26,7 @@ export const renderArrayType = ({
   renderer: (props: JsonSchemaFormProps) => React.ReactNode;
   onChange: (value: SorobanInvokeValue) => void;
 }) => {
-  const name = path.join("-");
+  const name = path.join(".");
   const invokeContractBaseProps = {
     contract_id: parsedSorobanOperation.contract_id,
     function_name: parsedSorobanOperation.function_name,
@@ -41,6 +35,25 @@ export const renderArrayType = ({
   const schemaItems = jsonSchema.getSchemaItems(schema);
 
   const nestedArgsItems = get(parsedSorobanOperation.args, name) || [];
+
+  const disableAddButton =
+    jsonSchema.isTuple(schema) ||
+    schemaItems.maxItems === nestedArgsItems.length;
+
+  if (jsonSchema.isTuple(schema)) {
+    return schemaItems.map((item: JSONSchema7, index: number) => {
+      // will render primitive/array/all the types
+      const nestedPath = [name, index].join(".");
+
+      return renderer({
+        name: nestedPath,
+        schema: item,
+        path: [nestedPath],
+        parsedSorobanOperation,
+        onChange,
+      });
+    });
+  }
 
   return (
     <Box gap="md" key={`${name}`}>
@@ -65,7 +78,19 @@ export const renderArrayType = ({
                   {/* Map Type (scSpecTypeMap) */}
                   {jsonSchema.isSchemaObject(schema.items) &&
                   schema.items.type === "object" ? (
-                    <LabelHeading size="lg">{nestedPathTitle}</LabelHeading>
+                    // <LabelHeading size="lg">{nestedPathTitle}sss</LabelHeading>
+
+                    Object.keys(args).map((arg) => {
+                      const nestedPath = [index, arg].join(".");
+
+                      return renderer({
+                        name: nestedPathTitle,
+                        schema: schemaItems?.[arg] as JSONSchema7,
+                        path: [...path, nestedPath],
+                        parsedSorobanOperation,
+                        onChange,
+                      });
+                    })
                   ) : (
                     <Box gap="md">
                       {/* Vec Type (scSpecTypeVec) */}
@@ -85,18 +110,16 @@ export const renderArrayType = ({
           );
         })}
 
-      {!jsonSchema.isTuple(schema) ? (
+      {!disableAddButton ? (
         <Box gap="md" direction="row" align="center">
           <Button
             variant="secondary"
             size="md"
             onClick={() => {
-              const template = getTemplate({ schemaItems });
+              const template = getTemplate({ schema });
 
               const args = parsedSorobanOperation.args?.[name] || [];
               args.push(template);
-
-              console.log("[addItem] args: ", args);
 
               onChange({
                 ...invokeContractBaseProps,
@@ -116,47 +139,30 @@ export const renderArrayType = ({
   );
 };
 
-const getTemplate = ({ schemaItems }: { schemaItems: JSONSchema7 }) => {
+const getTemplate = ({ schema }: { schema: JSONSchema7 }) => {
   let template: AnyObject = {};
 
-  console.log("[addItem] schemaItems: ", schemaItems);
+  const schemaItems = jsonSchema.getSchemaItems(schema);
 
-  // Map Type (scSpecTypeMap)
-  if (schemaItems.type === "array" && schemaItems.items) {
-    // {
-    //   "items": {
-    //     "type": "object",
-    //     "properties": {
-    //       "address": { "type": "string" },
-    //       "amount": { "type": "string" },
-    //       "request_type": { "type": "string" }
-    //     }
-    //   }
-    // }
-    // then the defaultTemplate will be:
-    // {
-    //   "address": {},
-    //   "amount": {},
-    //   "request_type": {}
-    // }
+  if (
+    jsonSchema.isSchemaObject(schema.items) &&
+    schema.items.type === "object"
+  ) {
+    template = Object.keys(schemaItems || {}).reduce(
+      (acc: Record<string, string | AnyObject>, key) => {
+        // Good example:
+        // CDVQVKOY2YSXS2IC7KN6MNASSHPAO7UN2UR2ON4OI2SKMFJNVAMDX6DP
+        // submit function with object args
+        if (key === "additionalProperties") {
+          return acc;
+        }
+
+        acc[key] = {};
+        return acc;
+      },
+      {},
+    );
   }
 
   return template;
-
-  // Vec Type (scSpecTypeVec)
-  // Array of simple element types
-  //   if (schemaItems) {
-  //     console.log("[schemaType array] schemaItems: ", schemaItems);
-
-  //     return <div>Array</div>;
-  //   }
-
-  //   switch (schemaType) {
-  //     case "tuple":
-  //       return <div>Object</div>;
-  //     case "map":
-  //       return <div>Array</div>;
-  //     default:
-  //       return <div>Default</div>;
-  //   }
 };
